requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
vaderSentiment==3.3.2
textblob==0.17.1
influxdb-client==1.38.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
schedule==1.2.0
smtplib-ssl==1.0.4
python-dotenv==1.0.0
logging==0.4.9.6
scipy==1.11.4
plotly==5.17.0
kaleido==0.2.1